modified:   data/battle_scripts_1.s 	modified:   data/layouts/Elite4_Hoenn_2/map.bin 	modified:   data/layouts/Hoenn2_4/map.bin 	modified:   data/layouts/Hoenn3_6/map.bin 	modified:   data/layouts/layouts.json 	modified:   data/maps/Elite4_Hoenn_1/map.json 	modified:   data/maps/Elite4_Hoenn_1/scripts.inc 	new file:   data/maps/Elite4_Hoenn_1/scripts.pory 	modified:   data/maps/Elite4_Hoenn_2/map.json 	modified:   data/maps/Elite4_Hoenn_2/scripts.inc 	new file:   data/maps/Elite4_Hoenn_2/scripts.pory 	modified:   data/maps/Elite4_Hoenn_Outside/scripts.inc 	new file:   data/maps/Elite4_Hoenn_Outside/scripts.pory 	modified:   data/maps/Hoenn1_1/map.json 	modified:   data/maps/Hoenn1_1/scripts.inc 	modified:   data/maps/Hoenn1_1/scripts.pory 	modified:   data/maps/Hoenn1_3/scripts.inc 	modified:   data/maps/Hoenn1_3/scripts.pory 	modified:   data/maps/Hoenn1_4/scripts.inc 	modified:   data/maps/Hoenn1_4/scripts.pory 	modified:   data/maps/Hoenn1_5/scripts.inc 	modified:   data/maps/Hoenn1_5/scripts.pory 	modified:   data/maps/Hoenn1_6/scripts.inc 	modified:   data/maps/Hoenn1_6/scripts.pory 	modified:   data/maps/Hoenn2_1/map.json 	modified:   data/maps/Hoenn2_1/scripts.inc 	modified:   data/maps/Hoenn2_1/scripts.pory 	modified:   data/maps/Hoenn2_4/map.json 	modified:   data/maps/Hoenn2_4/scripts.inc 	modified:   data/maps/Hoenn2_4/scripts.pory 	modified:   data/maps/Hoenn3_1/map.json 	modified:   data/maps/Hoenn3_1/scripts.inc 	modified:   data/maps/Hoenn3_1/scripts.pory 	modified:   data/maps/Hoenn3_2/map.json 	modified:   data/maps/Hoenn3_2/scripts.inc 	modified:   data/maps/Hoenn3_2/scripts.pory 	modified:   data/maps/Hoenn3_5/map.json 	modified:   data/maps/Hoenn3_5/scripts.inc 	modified:   data/maps/Hoenn3_5/scripts.pory 	modified:   data/maps/Hoenn3_6/map.json 	modified:   data/maps/Hoenn3_6/scripts.inc 	modified:   data/maps/Hoenn3_6/scripts.pory 	modified:   data/maps/Hub1/scripts.inc 	modified:   data/maps/Hub1/scripts.pory 	modified:   data/maps/Route109/scripts.inc 	modified:   include/battle_scripts.h 	modified:   include/constants/abilities.h 	modified:   include/constants/battle_string_ids.h 	modified:   include/constants/opponents.h 	modified:   include/constants/opponents_frlg.h 	modified:   include/constants/party_menu.h 	modified:   include/party_menu.h 	modified:   include/pokemon_summary_screen.h 	modified:   include/starter_choose.h 	modified:   src/battle_ai_switch.c 	modified:   src/battle_ai_util.c 	modified:   src/battle_anim_effects_3.c 	modified:   src/battle_controller_player.c 	modified:   src/battle_main.c 	modified:   src/battle_message.c 	modified:   src/battle_move_resolution.c 	modified:   src/battle_script_commands.c 	modified:   src/battle_setup.c 	modified:   src/battle_switch_in.c 	modified:   src/battle_util.c 	modified:   src/battle_util2.c 	modified:   src/data/abilities.h 	modified:   src/data/choose_random_pool.h 	modified:   src/data/choose_random_pool_all_raw.h 	modified:   src/data/items.h 	modified:   src/data/moves_info.h 	modified:   src/data/pokemon/species_info/gen_1_families.h 	new file:   src/data/pokemon/species_info/gen_1_families.h.bak 	modified:   src/data/pokemon/species_info/gen_2_families.h 	new file:   src/data/pokemon/species_info/gen_2_families.h.bak 	modified:   src/data/pokemon/species_info/gen_3_families.h 	new file:   src/data/pokemon/species_info/gen_3_families.h.bak 	modified:   src/data/pokemon/species_info/gen_4_families.h 	new file:   src/data/pokemon/species_info/gen_4_families.h.bak 	modified:   src/data/pokemon/species_info/gen_5_families.h 	modified:   src/data/pokemon/species_info/gen_6_families.h 	modified:   src/data/pokemon/species_info/gen_7_families.h 	modified:   src/data/pokemon/species_info/gen_8_families.h 	modified:   src/data/pokemon/species_info/gen_9_families.h 	modified:   src/data/trainers.party 	modified:   src/data/trainers_frlg.party 	modified:   src/item_menu.c 	modified:   src/item_use.c 	modified:   src/overworld.c 	modified:   src/party_menu.c 	modified:   src/pokemon.c 	modified:   src/pokemon_summary_screen.c 	modified:   src/starter_choose.c 	modified:   test/battle/ability/early_bird.c 	new file:   test/battle/ability/fire_mane.c 	modified:   test/battle/ai/check_bad_move.c 	new file:   tools/species-editor/.gitignore 	new file:   tools/species-editor/README.md 	new file:   tools/species-editor/SpeciesEditor.spec 	new file:   tools/species-editor/build_exe.ps1 	new file:   tools/species-editor/build_exe_noport.ps1 	new file:   tools/species-editor/desktop_launcher.py 	new file:   tools/species-editor/desktop_noport.py 	new file:   tools/species-editor/public/index.html 	new file:   tools/species-editor/public/styles.css 	new file:   tools/species-editor/server.py

+++ include/starter_choose.h
@@ -16,7 +16,7 @@ struct BirchBagWeightedChoice
 
 u16 GetStarterPokemon(u16 chosenStarterId);
 void CB2_ChooseStarter(void);
-void SetBirchBagWeightedChoices(const struct BirchBagWeightedChoice *choices, u8 count);
+void SetBirchBagWeightedChoices(const struct BirchBagWeightedChoice *choices, u16 count);
 void ResetBirchBagWeightedPool(void);
 void ChooseBirchBagPokemonWeighted(void);
